Split test cases for BLAS operators
