+++ tests/test_mv.py
@@ -1,0 +1,36 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    MN_SHAPES = [
+        (1, 32),
+    ]
+    FLOAT_DTYPES = [torch.float32]
+else:
+    MN_SHAPES = [
+        (1, 32),
+        (160, 1024),
+        (5333, 497),
+    ]
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+
+
+@pytest.mark.mv
+@pytest.mark.parametrize("M, N", MN_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_mv(M, N, dtype):
+    matrix = torch.randn((N, M), dtype=dtype, device=flag_gems.device)
+    vector = torch.randn((M,), dtype=dtype, device=flag_gems.device)
+    ref_matrix = utils.to_reference(matrix, True)
+    ref_vector = utils.to_reference(vector, True)
+
+    ref_out = torch.mv(ref_matrix, ref_vector)
+    with flag_gems.use_gems():
+        res_out = torch.mv(matrix, vector)
+
+    utils.gems_assert_close(res_out, ref_out, dtype, reduce_dim=M)
