+++ tests/test_outer.py
@@ -1,0 +1,46 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    MN_SHAPES = [
+        (1, 32),
+    ]
+    FLOAT_DTYPES = [torch.float32]
+else:
+    MN_SHAPES = [
+        (1, 32),
+        (160, 1024),
+        (5333, 497),
+    ]
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+
+
+@pytest.mark.outer
+@pytest.mark.parametrize(
+    "M, N", MN_SHAPES + ([(32, 131072)] if flag_gems.vendor_name == "cambricon" else [])
+)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_outer(M, N, dtype):
+    inp1 = torch.randn(M, dtype=dtype, device=flag_gems.device, requires_grad=True)
+    inp2 = torch.randn(N, dtype=dtype, device=flag_gems.device, requires_grad=True)
+    ref_inp1 = utils.to_reference(inp1, True)
+    ref_inp2 = utils.to_reference(inp2, True)
+
+    ref_out = torch.outer(ref_inp1, ref_inp2)
+    res_out = flag_gems.outer(inp1, inp2)
+    utils.gems_assert_close(res_out, ref_out, dtype)
+
+    out_grad = torch.randn_like(res_out)
+    ref_grad = utils.to_reference(out_grad, True)
+
+    ref_in1_grad, ref_in2_grad = torch.autograd.grad(
+        ref_out, (ref_inp1, ref_inp2), ref_grad
+    )
+    res_in1_grad, res_in2_grad = torch.autograd.grad(res_out, (inp1, inp2), out_grad)
+    utils.gems_assert_close(res_in1_grad, ref_in1_grad, dtype, reduce_dim=N)
+    utils.gems_assert_close(res_in2_grad, ref_in2_grad, dtype, reduce_dim=M)
